+++ src/main/java/fr/maxlego08/essentials/user/ZUser.java
@@ -1,6 +1,6 @@
 package fr.maxlego08.essentials.user;
 
-import com.tcoded.folialib.impl.ServerImplementation;
+import com.tcoded.folialib.impl.PlatformScheduler;
 import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.Permission;
 import fr.maxlego08.essentials.api.dto.CooldownDTO;
@@ -231,13 +231,13 @@ public class ZUser extends ZUtils implements User {
         Location playerLocation = getPlayer().getLocation();
         AtomicInteger atomicInteger = new AtomicInteger(teleportationModule.getTeleportationDelay(getPlayer()));
 
-        if (teleportationModule.isTeleportDelayBypass() && this.hasPermission(Permission.ESSENTIALS_TELEPORT_BYPASS)) {
+        if (teleportationModule.isTeleportDelayBypass() && this.hasPermission(Permission.ESSENTIALS_TELEPORT_BYPASS) || atomicInteger.get() <= 0) {
             this.teleport(teleportationModule, location, successMessage, args);
             return;
         }
 
-        ServerImplementation serverImplementation = this.plugin.getScheduler();
-        serverImplementation.runAtLocationTimer(location, wrappedTask -> {
+        PlatformScheduler platformScheduler = this.plugin.getScheduler();
+        platformScheduler.runAtLocationTimer(location, wrappedTask -> {
 
             if (!this.isOnline()) {
                 wrappedTask.cancel();